fixed Storage id bug

+++ main.go
@@ -62,7 +62,7 @@ func FetchStorages(dev *mtp.Device) ([]StorageData, error) {
 
 	for _, sid := range sids.Values {
 		var info mtp.StorageInfo
-		if err := dev.GetStorageInfo(uint32(sid), &info); err != nil {
+		if err := dev.GetStorageInfo(sid, &info); err != nil {
 			return nil, StorageInfoError{error: err}
 		}
